refactor(skaha): consolidate metrics package into fewer types
Collapse platform and pod mappers into PlatformMetrics and PodMetrics,
move test fixtures to test sources, and keep eight production metrics types.
