--- skaha/src/main/java/org/opencadc/skaha/metrics/PlatformMetricsMetadata.java
@@ -1,10 +1,0 @@
-package org.opencadc.skaha.metrics;
-
-import java.time.Instant;
-
-/**
- * Metrics response metadata for a platform metrics snapshot.
- *
- * @param created snapshot time ({@code metadata.created} from the Metrics API)
- */
-public record PlatformMetricsMetadata(Instant created) {}
