+++ skaha/src/main/java/org/opencadc/skaha/session/SessionDAO.java
@@ -20,7 +20,7 @@ import org.opencadc.skaha.K8SUtil;
 import org.opencadc.skaha.KubernetesJob;
 import org.opencadc.skaha.SkahaAction;
 import org.opencadc.skaha.metrics.MetricsDAO;
-import org.opencadc.skaha.metrics.PodMetricsMapper;
+import org.opencadc.skaha.metrics.PodMetrics;
 import org.opencadc.skaha.metrics.PodResourceUsage;
 import org.opencadc.skaha.metrics.SkahaMetricsDAO;
 
@@ -180,7 +180,7 @@ public class SessionDAO {
     static PodResourceUsage loadPodResourceUsage(
             final MetricsDAO metricsDAO, final String forUserID, final boolean omitHeadless) {
         try {
-            return PodMetricsMapper.toPodResourceUsage(metricsDAO.getPodMetrics(forUserID, omitHeadless));
+            return PodMetrics.toPodResourceUsage(metricsDAO.getPodMetrics(forUserID, omitHeadless));
         } catch (Exception e) {
             LOGGER.warn("Failed to fetch pod metrics for sessions: " + e.getMessage(), e);
             return PodResourceUsage.empty();
